Refactor to move functions that relate to specfic chart types to the relevant classes
